Replace render props with simply rendering children components
