+++ static/js/publisher/release/components/releasesTableRevisionsRow.js
@@ -2,24 +2,26 @@ import React from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
 
+import { getArchitectures } from "../selectors";
+import { isSameVersion } from "../helpers";
 import { ReleasesTableRevisionCell } from "./releasesTableCell";
-
 import ReleasesTableRow from "./releasesTableRow";
 
 const ReleasesTableRevisionsRow = props => {
-  const { currentTrack, risk, revisions } = props;
+  const { currentTrack, risk, revisions, archs } = props;
+  const showVersion = !isSameVersion(revisions);
 
   return (
     <ReleasesTableRow risk={risk} revisions={revisions}>
-      {({ arch, hasSameVersion }) => {
+      {archs.map(arch => {
         return (
           <ReleasesTableRevisionCell
             key={`${currentTrack}/${risk}/${arch}`}
             revision={revisions ? revisions[arch] : null}
-            showVersion={!hasSameVersion}
+            showVersion={showVersion}
           />
         );
-      }}
+      })}
     </ReleasesTableRow>
   );
 };
@@ -30,12 +32,14 @@ ReleasesTableRevisionsRow.propTypes = {
   revisions: PropTypes.object,
 
   // state
-  currentTrack: PropTypes.string.isRequired
+  currentTrack: PropTypes.string.isRequired,
+  archs: PropTypes.array.isRequired
 };
 
 const mapStateToProps = state => {
   return {
-    currentTrack: state.currentTrack
+    currentTrack: state.currentTrack,
+    archs: getArchitectures(state)
   };
 };
